src/services/master_agent/requirements.txt
fastapi
uvicorn
httpx
groq
python-dotenv
pydantic